move get_detections to evaluate.py
