+++ evaluate.py
@@ -5,8 +5,11 @@ import os.path as osp
 import numpy as np
 import torch
 from torchvision import transforms
+
 from tqdm import tqdm
 
+from models.utils import get_bboxes
+from utils.nms import nms
 import trainer
 from datasets import get_dataloader
 from models.model import DetectionModel
@@ -51,6 +54,72 @@ def get_model(checkpoint=None, num_templates=25):
     return model
 
 
+def get_detections(model,
+                   img,
+                   templates,
+                   rf,
+                   img_transforms,
+                   prob_thresh=0.65,
+                   nms_thresh=0.3,
+                   scales=(-2, -1, 0, 1),
+                   device=None):
+    model = model.to(device)
+    model.eval()
+
+    dets = np.empty((0, 5))  # store bbox (x1, y1, x2, y2), score
+
+    num_templates = templates.shape[0]
+
+    # Evaluate over multiple scale
+    scales_list = [2**x for x in scales]
+
+    # convert tensor to PIL image so we can perform resizing
+    image = transforms.functional.to_pil_image(img[0])
+
+    min_side = np.min(image.size)
+
+    for scale in scales_list:
+        # scale the images
+        scaled_image = transforms.functional.resize(image,
+                                                    np.int(min_side * scale))
+
+        # normalize the images
+        img = img_transforms(scaled_image)
+
+        # add batch dimension
+        img.unsqueeze_(0)
+
+        # now run the model
+        x = img.float().to(device)
+
+        output = model(x)
+
+        # first `num_templates` channels are class maps
+        score_cls = output[:, :num_templates, :, :]
+        prob_cls = torch.sigmoid(score_cls)
+
+        score_cls = score_cls.data.cpu().numpy().transpose((0, 2, 3, 1))
+        prob_cls = prob_cls.data.cpu().numpy().transpose((0, 2, 3, 1))
+
+        score_reg = output[:, num_templates:, :, :]
+        score_reg = score_reg.data.cpu().numpy().transpose((0, 2, 3, 1))
+
+        t_bboxes, scores = get_bboxes(score_cls, score_reg, prob_cls,
+                                      templates, prob_thresh, rf, scale)
+
+        scales = np.ones((t_bboxes.shape[0], 1)) / scale
+        # append scores at the end for NMS
+        d = np.hstack((t_bboxes, scores))
+
+        dets = np.vstack((dets, d))
+
+    # Apply NMS
+    keep = nms(dets, nms_thresh)
+    dets = dets[keep]
+
+    return dets
+
+
 def write_results(dets, img_path, split, results_dir=None):
     results_dir = results_dir or "{0}_results".format(split)
 
@@ -87,14 +156,14 @@ def run(model,
         debug=False):
     for _, (img, filename) in tqdm(enumerate(val_loader),
                                    total=len(val_loader)):
-        dets = trainer.get_detections(model,
-                                      img,
-                                      templates,
-                                      val_loader.dataset.rf,
-                                      val_loader.dataset.transforms,
-                                      prob_thresh,
-                                      nms_thresh,
-                                      device=device)
+        dets = get_detections(model,
+                              img,
+                              templates,
+                              val_loader.dataset.rf,
+                              val_loader.dataset.transforms,
+                              prob_thresh,
+                              nms_thresh,
+                              device=device)
 
         write_results(dets, filename[0], split, results_dir)
     return dets
